fixed immediate linker problem
