+++ FinalProject/FinalProject/FinalProject.vcxproj.filters
@@ -24,6 +24,9 @@
     <ClCompile Include="ReservationSystem.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Flight.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Passenger.h">
